requirements.txt
discord.py
steam
